Fix panic on long sequence numbers
When /proc/net/tcp gets big due to the number of TCP sockets,
the sequence numbers start using up the space they are aligned
with, therefore the regular expression "\s+" is no longer
sufficient to parse the lines. Luckily, there's a ":" after the
number, so let's split the line there first.

I have a rule to introduce testing in a project when the first
bug appears, so I've already wrote the first tests to fix this.