Detecting directionality
This change enables the program to guess directionality based on:
- ports listening on specific IPs
- ports listening on 0.0.0.0
- ports that are ephemeral on the system it's running on

The latter two are not exactly reliable unfortunately, since what's
an outgoing port on one node, could be a listening port on another.

The best solution would be to parse out all IP addresses assigned
to the current node from /proc/net/fib_trie, but that's not that
simple and I probably won't have time for that.