Implemented event-based flow
Refactored the whole program to use events instead of the good
old imperative flow.

This has a ton of advantages, just a few examples:
- Tests are becoming much simpler, since the different parts
  that have different concerns are loosely coupled.
- It's way easier to add new functionality
- Multiple sources can be added independently from each other

Of course the disadvantage here is that an event-based system
is a bit harder to debug sometimes, but the complexity of this
code is far from that.